+++ src/Collections/AbstractCollection.php
@@ -5,7 +5,10 @@ namespace GingerPluginSdk\Collections;
 
 use GingerPluginSdk\Interfaces\MultiFieldsEntityInterface;
 
-/** @phpstan-template T */
+/**
+ * @template T
+ * @phpstan-template T
+ */
 class AbstractCollection implements MultiFieldsEntityInterface
 {
     private int $pointer = 0;
@@ -36,20 +39,34 @@ class AbstractCollection implements MultiFieldsEntityInterface
         return $this;
     }
 
-    /** @phpstan-param T $item */
+    /**
+     * @param T $item
+     *
+     * @phpstan-param T $item
+     */
     public function add(mixed $item): void
     {
         $this->next();
         $this->items[$this->pointer] = $item;
     }
 
-    /** @return T|null */
+    /**
+     * @param int|string $position
+     * @return T|null
+     *
+     * @phpstan-param int|string $position
+     * @phpstan-return T|null
+     */
     public function get($position = null)
     {
         return $this->items[$position ?? $this->pointer];
     }
 
-    /** @return T[] */
+    /**
+     * @return T[]
+     *
+     * @phpstan-return T[]
+     */
     public function getAll(): array
     {
         return $this->items;